Support details with scale=0, double-precision AABB

+++ Mlib/Render/Render_Logics/Imposter_Logic.hpp
@@ -91,7 +91,7 @@ private:
     float down_sampling_;
     float max_deviation_;
     float min_distance_;
-    AxisAlignedBoundingBox<float, 3> obj_relative_aabb_;
+    AxisAlignedBoundingBox<double, 3> obj_relative_aabb_;
 };
 
 }
